+++ xv6-riscv/user/usys.pl
@@ -46,3 +46,5 @@ entry("setpriority");
 entry("getpriority");
 entry("jail");
 entry("agent_recv");
+entry("set_deny");
+entry("get_deny");